Extract CanvasProvider module and share it across ScalerContext and PathRasterizer.
