+++ web/src/wechat/binding.ts
@@ -21,15 +21,16 @@ import { setTGFXModule } from '../tgfx-module';
 import * as tgfx from './tgfx';
 import { Matrix } from '../core/matrix';
 import { ScalerContext } from '../core/scaler-context';
-import { PathRasterizer } from './path-rasterizer';
+import { PathRasterizer } from '../core/path-rasterizer';
+import { setCanvasProvider } from '../core/canvas-provider';
 import { getCanvas2D, releaseCanvas2D } from './canvas';
 
 export const TGFXBind = (module: TGFX) => {
   setTGFXModule(module);
   module.module = module;
-  // Inject the WeChat-specific canvas provider so that ScalerContext reuses the
+  // Inject the WeChat-specific canvas provider so that core rasterizers reuse the
   // wx.createOffscreenCanvas-based pool instead of the browser canvas utilities.
-  ScalerContext.setCanvasProvider({
+  setCanvasProvider({
     getCanvas2D,
     releaseCanvas2D: releaseCanvas2D as (canvas: HTMLCanvasElement | OffscreenCanvas) => void,
   });
